+++ AffirmSDK/AffirmBaseModalViewController.m
@@ -14,8 +14,11 @@
 
 - (instancetype)init {
     if (self = [super init]) {
-        self.webView = [UIWebView new];
-        self.webView.delegate = self;
+        self.webView = [[WKWebView alloc] init];
+        self.webView.navigationDelegate = self;
+        self.webView.UIDelegate = self;
+        
+        self.loadingIndicator = [AffirmActivityIndicator new];
     }
     return self;
 }
@@ -26,51 +29,45 @@
     [self renderWebView];
 }
 
-- (void)viewDidAppear:(BOOL)animated {
-    [super viewDidAppear:animated];
-    [self showLoadingIndicator];
-}
-
 - (void)renderWebView {
     [self.view addSubview:self.webView];
     self.webView.translatesAutoresizingMaskIntoConstraints = NO;
     self.webView.scrollView.bounces = NO;
+    [self.webView addObserver:self forKeyPath:NSStringFromSelector(@selector(estimatedProgress)) options:NSKeyValueObservingOptionNew context:NULL];
     NSLayoutConstraint * top = [NSLayoutConstraint constraintWithItem:self.webView attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.topLayoutGuide attribute:NSLayoutAttributeBottom multiplier:1 constant:0];
     NSLayoutConstraint * bottom = [NSLayoutConstraint constraintWithItem:self.webView attribute:NSLayoutAttributeBottom relatedBy:NSLayoutRelationEqual toItem:self.bottomLayoutGuide attribute:NSLayoutAttributeTop multiplier:1 constant:0];
     NSLayoutConstraint * left = [NSLayoutConstraint constraintWithItem:self.webView attribute:NSLayoutAttributeLeft relatedBy:NSLayoutRelationEqual toItem:self.view attribute:NSLayoutAttributeLeft multiplier:1 constant:0];
     NSLayoutConstraint * right = [NSLayoutConstraint constraintWithItem:self.webView attribute:NSLayoutAttributeRight relatedBy:NSLayoutRelationEqual toItem:self.view attribute:NSLayoutAttributeRight multiplier:1 constant:0];
     [self.view addConstraints:@[top, bottom, left, right]];
     
-    self.loadingIndicator = [AffirmActivityIndicator new];
     self.loadingIndicator.center = self.view.center;
     self.loadingIndicator.indicatorColor = [UIColor colorWithRed:16/255.0 green:160/255.0 blue:234/255.0 alpha: 1.0];
     self.loadingIndicator.indicatorPathColor = [UIColor colorWithRed:246/255.0 green:248/255.0 blue:252/255.0 alpha:1.0];
     self.loadingIndicator.indicatorLineWidth = 1.5;
+    [self.loadingIndicator startAnimatingOnView:self.view];
+}
+
+- (void)dealloc {
+    [self.webView removeObserver:self forKeyPath:NSStringFromSelector(@selector(estimatedProgress))];
+    self.webView.navigationDelegate = nil;
+    self.webView.UIDelegate = nil;
 }
 
 #pragma mark - Webview
 
-- (void) showLoadingIndicator {
-    if (self.loadingCount == 0) {
-        [self.loadingIndicator startAnimatingOnView:self.view];
-    }
-    self.loadingCount += 1;
-}
-
-- (void) hideLoadingIndicator {
-    self.loadingCount -= 1;
-    if (self.loadingCount <= 0) {
-        [self.loadingIndicator stopAnimating];
+- (void)observeValueForKeyPath:(NSString *)keyPath ofObject:(id)object change:(NSDictionary<NSKeyValueChangeKey,id> *)change context:(void *)context {
+    if ([keyPath isEqualToString:NSStringFromSelector(@selector(estimatedProgress))] && object == self.webView) {
+        if (self.webView.estimatedProgress >= 1.0f) {
+            [self.loadingIndicator stopAnimating];
+        }
+    } else {
+        [super observeValueForKeyPath:keyPath ofObject:object change:change context:context];
     }
 }
 
-- (void) webViewDidFinishLoad:(UIWebView *)webView {
-    [self hideLoadingIndicator];
-}
-
-- (void) webView:(UIWebView *)webView didFailLoadWithError:(NSError *)error {
-    [self hideLoadingIndicator];
-    [AffirmLogger logEvent:@"Web load failed" info:@{@"failed_url": webView.request.URL.absoluteString, @"error_description" : error.localizedDescription}];
+- (void) webView:(WKWebView *)webView didFailNavigation:(WKNavigation *)navigation withError:(NSError *)error {
+    [self.loadingIndicator stopAnimating];
+    [AffirmLogger logEvent:@"Web load failed" info:@{@"error_description" : error.localizedDescription}];
     NSString *url = [NSString stringWithFormat:@"https://www.affirm.com/u/#/error?main=Error&sub=%@", [error.localizedDescription stringByAddingPercentEncodingWithAllowedCharacters:[NSCharacterSet URLQueryAllowedCharacterSet]]];
     [self.webView loadRequest:[NSURLRequest requestWithURL:[NSURL URLWithString:url]]];
 }